Tripletex AI Accounting Agent/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
requests==2.32.3
google-cloud-aiplatform>=1.71.0
google-cloud-storage>=2.18.0
google-genai>=1.0.0
pydantic==2.9.0
python-multipart==0.0.9
gunicorn==22.0.0
